Converted menu loop to new MenuOption array implementation
Menu loop now consists of only two function calls within a loop.
The work of displaying the menu, parsing the input, and executing 
the required function are now handled by these two functions:

char displayMenuAndReturnOption(struct MenuOption menuArray[]);
int executeMenuOption(struct MenuOption menuArray[], 
char employeeArray[][50], char menuOptChar);

Although there is a little complexity in the new functions, what is 
important is they don't need to change as the function grows.  
In other words, there are no configurable options or data within 
these functions, because they are all concentrated in the MenuOptions array.

Consequently, while the previous implementation spread the 
menu messages, the menu key, and the menu code across pages of 
code, the new implementation puts this all in one line of code per option.
Arguable, the only new bit of complexity is that a function definition is 
now required for each new menu option, but since this is an opportunity 
to provide another informative label, this is really an aid to keeping the 
menu description high level without getting bogged down in implementation.